Improve readme and manual; move plugin-related code to it's own module
